add cljsjs/d3 dependency
Fairly big changes between D3 `4.0.0` (which we've required here) and
`3.5.17`, which is used in the Untangled Tutorial and (likely also) in
the Sunburst example we're trying to port/replicate. So, we may want to
consider rolling back to 3.5.x if we hit any major snags.

D3 changelog - https://github.com/d3/d3/blob/master/CHANGES.md